Not a number (NaN) handling in every brick action:
* Formula interpretation to NaN will be handled + test.
* Null-Formula tests.
* Stringfunctionhandling of Not-A-Number parameters.

+++ catroidTest/src/org/catrobat/catroid/test/io/StorageHandlerTest.java
@@ -132,9 +132,11 @@ public class StorageHandlerTest extends AndroidTestCase {
 		assertEquals("Size was not deserialized right", size,
 				actualSize.interpretFloat(postSpriteList.get(1).getScript(0).getBrickList().get(2).getSprite()));
 		assertEquals("XPosition was not deserialized right", xPosition,
-				actualXPosition.interpretInteger(postSpriteList.get(2).getScript(0).getBrickList().get(0).getSprite()));
+				actualXPosition.interpretInteger(postSpriteList.get(2).getScript(0).getBrickList().get(0).getSprite())
+						.intValue());
 		assertEquals("YPosition was not deserialized right", yPosition,
-				actualYPosition.interpretInteger(postSpriteList.get(2).getScript(0).getBrickList().get(0).getSprite()));
+				actualYPosition.interpretInteger(postSpriteList.get(2).getScript(0).getBrickList().get(0).getSprite())
+						.intValue());
 
 		assertFalse("paused should not be set in script", preSpriteList.get(1).getScript(0).isPaused());
